test: consolidate cwd alias coverage
